Create photogen_osCommerce.sql
This is the database structure required for the usage of my system. Of course it can be easily edited. 

If the names and rows of the db is changed, simply just edit connect.php file as all the names of the rows/columns are store in the variable in connect.php to allow easy access across the entire file.

So you won't have to keep editing one by one through the entire file.